modified migrations, models, and seeders

+++ database/migrations/2025_03_29_233911_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->string('code')->unique();
-            $table->bigInteger('price');
+            $table->unsignedInteger('price');
             $table->timestamps();
         });
     }
